+++ src/main/java/kitchenpos/menu/domain/MenuGroup.java
@@ -1,4 +1,6 @@
-package kitchenpos.domain;
+package kitchenpos.menu.domain;
+
+import kitchenpos.domain.Name;
 
 import javax.persistence.*;
