+++ src/main/java/kitchenpos/menu/domain/MenuProducts.java
@@ -1,4 +1,4 @@
-package kitchenpos.domain;
+package kitchenpos.menu.domain;
 
 import javax.persistence.CascadeType;
 import javax.persistence.Embeddable;
