+++ src/main/java/kitchenpos/menu/dto/MenuGroupRequest.java
@@ -1,6 +1,6 @@
-package kitchenpos.dto.request;
+package kitchenpos.menu.dto;
 
-import kitchenpos.domain.MenuGroup;
+import kitchenpos.menu.domain.MenuGroup;
 
 public class MenuGroupRequest {
