+++ src/main/java/kitchenpos/menu/dto/MenuProductResponse.java
@@ -1,6 +1,6 @@
-package kitchenpos.dto.response;
+package kitchenpos.menu.dto;
 
-import kitchenpos.domain.MenuProduct;
+import kitchenpos.menu.domain.MenuProduct;
 
 import java.util.List;
 import java.util.stream.Collectors;
