+++ src/main/java/kitchenpos/menu/dto/MenuRequest.java
@@ -1,6 +1,6 @@
-package kitchenpos.dto.request;
+package kitchenpos.menu.dto;
 
-import kitchenpos.domain.Menu;
+import kitchenpos.menu.domain.Menu;
 
 import java.math.BigDecimal;
 import java.util.List;
