+++ src/main/java/kitchenpos/menu/repository/MenuGroupRepository.java
@@ -1,10 +1,7 @@
-package kitchenpos.dao;
+package kitchenpos.menu.repository;
 
-import kitchenpos.domain.MenuGroup;
+import kitchenpos.menu.domain.MenuGroup;
 import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.List;
-import java.util.Optional;
 
 public interface MenuGroupRepository extends JpaRepository<MenuGroup, Long> {
     boolean existsById(Long id);
