+++ src/main/java/kitchenpos/menu/repository/MenuRepository.java
@@ -1,10 +1,9 @@
-package kitchenpos.dao;
+package kitchenpos.menu.repository;
 
-import kitchenpos.domain.Menu;
+import kitchenpos.menu.domain.Menu;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
-import java.util.Optional;
 
 public interface MenuRepository extends JpaRepository<Menu, Long> {
